+++ ard/farm_aero/floris.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import numpy as np
 import floris
 
-import ard.utils.utils
+import ard.utils.core
 import ard.farm_aero.templates as templates
 
 
@@ -40,7 +40,7 @@ class FLORISFarmComponent:
         self.fmodel.set(
             wind_shear=self.modeling_options.get("wind_shear", 0.585),
             turbine_type=[
-                ard.utils.utils.create_FLORIS_turbine(self.modeling_options["turbine"])
+                ard.utils.core.create_FLORIS_turbine(self.modeling_options["turbine"])
             ],
         )
         self.fmodel.assign_hub_height_to_ref_height()
