+++ ard/utils/core.py
