+++ test/system/ard/LCOE_stack/test_LCOE_OFB_stack.py
@@ -9,7 +9,7 @@ from wisdem.optimization_drivers.nlopt_driver import NLoptDriver
 
 import ard
 import ard.utils.test_utils
-import ard.utils.utils
+import ard.utils.core
 import ard.wind_query as wq
 import ard.glue.prototype as glue
 import ard.cost.wisdem_wrap as cost_wisdem
@@ -35,7 +35,7 @@ class TestLCOE_OFB_stack:
             / "data"
             / "turbine_spec_IEA-22-284-RWT.yaml"
         )  # toolset generalized turbine specification
-        data_turbine_spec = ard.utils.utils.load_turbine_spec(filename_turbine_spec)
+        data_turbine_spec = ard.utils.core.load_turbine_spec(filename_turbine_spec)
 
         # set up the modeling options
         self.modeling_options = {
