+++ test/unit/ard/farm_aero/test_floris.py
@@ -5,7 +5,7 @@ import openmdao.api as om
 
 import floris
 
-import ard.utils.utils
+import ard.utils.core
 import ard.utils.test_utils
 import ard.wind_query as wq
 import ard.farm_aero.floris as farmaero_floris
@@ -42,7 +42,7 @@ class TestFLORISBatchPower:
             "data",
             "turbine_spec_IEA-3p4-130-RWT.yaml",
         ).absolute()  # toolset generalized turbine specification
-        data_turbine_spec = ard.utils.utils.load_turbine_spec(filename_turbine_spec)
+        data_turbine_spec = ard.utils.core.load_turbine_spec(filename_turbine_spec)
 
         # set up the modeling options
         modeling_options = {
@@ -151,7 +151,7 @@ class TestFLORISAEP:
             / "data"
             / "turbine_spec_IEA-3p4-130-RWT.yaml"
         )  # toolset generalized turbine specification
-        data_turbine_spec = ard.utils.utils.load_turbine_spec(filename_turbine_spec)
+        data_turbine_spec = ard.utils.core.load_turbine_spec(filename_turbine_spec)
 
         # set up the modeling options
         modeling_options = {
